requirements.txt
langchain
langchain-community
langchain-core
langchain-groq
langchain-google-community
langchain-nvidia-ai-endpoints
langchain-huggingface
python-dotenv
streamlit
pypdf
faiss-cpu
sentence-transformers
duckduckgo-search
arxiv
wikipedia
mysql-connector-python
psycopg2-binary
sqlalchemy
langchain-tavily